split app to pages
