+++ client/src/components/states/Loading.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export const Loading = () => {
+  return <div>Loading ...</div>;
+};
