+++ client/src/pages/Profile.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import { useAuth } from "../auth/useAuth";
+import { AddActivityForm } from "../components/forms/AddActivityForm";
+
+export const Profile = () => {
+  const { user } = useAuth();
+  return (
+    <>
+      <div>Welcome {user?.displayName}</div>
+      <AddActivityForm />
+    </>
+  );
+};
